Refactor Task into subclasses
