+++ src/lib/states/task-ecs.js
@@ -1,0 +1,7 @@
+const Task = require('./task');
+
+class TaskEcs extends Task {
+  // TODO: override abstract functions!
+}
+
+module.exports = TaskEcs;
